Removed NSQLookupDynMapImpl in favor of GSON-based implementation. Switched to Java7 source and target. Updated Maven POM file to refine dependencies scope. Added javadocs generation. Moved test files to src/test/java. Removed all binary files from version control. Removed all project files from version control.
