--- .classpath
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="lib/logback-classic-1.0.7.jar"/>
-	<classpathentry kind="lib" path="lib/logback-core-1.0.7.jar"/>
-	<classpathentry kind="lib" path="lib/netty-3.6.1.Final.jar"/>
-	<classpathentry kind="lib" path="lib/slf4j-api-1.6.4.jar"/>
-	<classpathentry kind="lib" path="lib/trendrr-oss.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
