+++ .gitignore
@@ -1,10 +1,13 @@
+# Generic IDE and binary excludes
 *.class
-
-build/classes
-build/jar
-build/build.num
+.idea
+*.iml
 target
-
-# Package Files #
-*.war
-*.ear
+*/target
+*/build
+*.pyc
+*.*~
+.classpath
+.project
+.settings
+.metadata
